requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]

### -e .